Add real device model to Analytics
Sourced from [1]. The code returns the actual device model, which we
have verified on a real device. Hard to test however as simulator
returns 'arm64' for an iPhone 15 pro.

It is not possible to fetch the full device model using UIDevice as
UIDevice.model and UIDevice.name only returns 'iPhone' for real devices.
Confusingly, running the app through simulator returns the full model
for UIDevice.name.

[1]: https://github.com/govuk-one-login/mobile-ios-networking/blob/00c9629759942f7464b8858965803030061348a1/Sources/Networking/HeaderTypes/DeviceModel.swift#L6.