[GOVUKAPP-1002] Add search autocomplete (#231)
* Add new API client for autocomplete

The autocomplete endpoint lives on the main gov.uk domain.

* Add GOVRequest for searchSuggestions

* Adds suggestions to search service

Autocomplete / search both originate from the same product and so
keeping specific error handling in SearchServiceClient seems reasonable
for future logging etc, even though we're not using error states past
the client for suggestions.

* Adds SearchSuggestionsViewModel

Fetches suggestions using search service. SearchSuggestion created with
a unique ID to force tableview to update everytime it reloads (important
for autocomplete suggestions feature).

* Adds suggestion highlighter

Requirements are to bolden words in the suggestion string that do not
match the passed in text. The highlighter should work for multiple
passed in words and match substrings individual words in the suggestion
string.

* Expose suggestion highlighter in view model

* Add suggestions view controller + cell

* Update suggestions tableview and cell

Aligns with search history for consistency.

Uses UIListContentConfiguration for cell config as much cleaner than
manually configuring.

* Update suggestions flow considering saved searches

Suggestions will appear on the third character inputted and dissapear if
under 3 chars (opposite for previous searches).

Suggestions will not appear if there aren't any to show.

Suggestions will hide if text is cleared.

* Limit suggestions to the first five

* Fix header not being focusable

* Remove Package changes

* Moves monitoring textfield editing to delegate

The delegate method is fired on return and thus have added a catch as we
don't want the suggestions block running on enter.

Also correctly returns in suggestions function in view model.

* Adds snapshot testing

Also fixes a couple issues:
- moves translatesAutoresizingMaskIntoConstraints flag to parent context
  (thanks Tom Bates)
- reloads snapshot on load