Reduce column count on topic onboarding
Reduces column count on topic onboarding for a given content size and
renders compact card which reduces the vertical footprint. This applies
for portrait and landscape when the content size is a given size.

It is a little crude sizing by solely content size but it seems to work
fine for smaller / larger iphones (se & iphone 16 pro).