Fixes background showing on scroll bounce
When scrolling to the top space emerges between the nav bar and title
due to the bounce. If you change the background colour to the nav blue,
the same thing happens at the bottom of the screen (blue shows).

Applying background gradient (first 1/3 blue) gives a resonable area to
allow the bounce without showing the background colour.