Combine AuthenticationTokenSet into AuthenticationService
Didn't really make sense for them to be seperate, especially as we'll
need to refresh the tokens regularly due to access token expiring after
60 seconds.

Also silently fails if encryption fails. This should be rare and will
just result in the user having to sign in again at a later point.