Add returning user logic to AuthenticationService
Handles new / returns users. If new user then core data and user
defaults are wiped and the new sub id is saved. Failure to save the new
sub id or wiping the data will result in a failure and force the user to
retry login again via an error screen.

Reloads SecureStoreService on signout which forces public / private key
generation if user attempts to log back in