[GOVUKAPP-1520] Handle different login users (#360)
* Add manager for persistent user identifier

The identifier will be used to differentiate between different users
logging into the app. It's saved using secure store but with an open
access level, as needs to be accessed from a logged out state.

* Add tests for PersistentUserIdentifierManager

Also fixes up tests.

* Add CoreData deletion service

Removes all data in CoreData, currently all objects in CoreData relate
to a user and need to be removed when signing in a new user.

* Add user defaults func to delete all

Deletes all keys defined in UserDefaultsKeys.

* Add returning user logic to AuthenticationService

Handles new / returns users. If new user then core data and user
defaults are wiped and the new sub id is saved. Failure to save the new
sub id or wiping the data will result in a failure and force the user to
retry login again via an error screen.

Reloads SecureStoreService on signout which forces public / private key
generation if user attempts to log back in

* Populates and uses localAuthenticationEnabled flag

Also moves this to localAuthenticationService

* Update user defaults mock

* Fetch topics remotely if empty

We wipe Topics objects if a new user signs in. Checks first if any local
and then fetches from remote if empty.

* New user onboarding

* Removes unused dismiss action from auth onboarding

* Remove unused let

* Fix test

* Fix topic onboarding

* Adds newUserAction + fixes tab setting

Utilises AppCoorindator as callback for NewUserOnboardingCoordinator
completion + sets home tab. Fixes two issues with tab not being set
after logging in and NewUserOnoboardingCoordinator being set as parent
of TabCoordinator leading to incorrect signout behaviour.

* Update email

If a user signs out and then straight back in again the email doesn't
get refreshed due to re-using the original tab controlle. Now updates
against the new idtoken on view appearance.

* Fixup tests

Setting email works when running the app but tests fail without being
MainActor / dispatched on main queue.

* Revert "Update email"

This reverts commit df60dd0f87b7a8fd731dcb660c14394cbafa8b7e.

* Revert "Fixup tests"

This reverts commit 5a3564e1150d2a35d1d575f0a4458792ce912561.

* Create new tab coordinator on signout

* Review fixes

* Small fixes