[GOVUKAPP-1682] Feature/inactivity auth (#372)
* Add inactivity service

Passing in a TimerWrapper to make testing easier.

Currently observes for activity:
- focus (single tap VO)
- keyboard strokes
- app becomes active (reset timer after biometrics scanned)

We'll monitor app touches in the SceneDelegate, which includes double
tap when using VO.

Also doesn't timeout users who are not signed in.

* Add tests for inactivity service

Mocking Timer made the tests a lot easier.

* Create custom UIWindow to monitor touch events

Resets inactivity timer on any touch event.

* Add inactivity coordinator

Also signs out when biometrics fails, as login screen is skipped if user
is signed in.

Reference to InactivityCoordinator created so it doesn't get removed
from memory.

Dismisses any modal that is presented explicitly otherwise the login
screen is set beneath an open modal.

* Fixup tests

* Add more tests

Moves GovUIWindow to own file. Adds tests for AppCoordinator,
InactivityCoordinator and GovUIWindow.

* Add InactivityCoordinator as child of AppCoordinator