Replaces AuthenticationOnboardingCoordinator
Adds new WelcomeOnboardingCoordinator. WelcomeOnboardingCoordinator
currently covers reauth / onboarding. We'll soon detach reauth and add
new coordinator for this case.

Removes use of onboarding module as designs slightly different and
AuthenticationOnboarding was the only feature to use the module. We
should reintroduce onboarding module when / if it can be applied to
multiple onboarding screens again.