[GOVUKAPP-1585] Feature/welcome onboarding (#379)
* Remove three slide onboarding screen

* Remove old onboarding assets + add new

* Remove unused slides code

* Replaces AuthenticationOnboardingCoordinator

Adds new WelcomeOnboardingCoordinator. WelcomeOnboardingCoordinator
currently covers reauth / onboarding. We'll soon detach reauth and add
new coordinator for this case.

Removes use of onboarding module as designs slightly different and
AuthenticationOnboarding was the only feature to use the module. We
should reintroduce onboarding module when / if it can be applied to
multiple onboarding screens again.

* Swaps local authority explainer image

The screen was using the default onboarding image, swaps to match
designs.

* Adds snapshot tests for welcome onboarding

* Rename function

* Add view model test

* Remove unused extension

* Consolidate welcome and auth info view

* Update snapshot tests

* Fix signin / out missing image

* Fix nested sonar warning