Adds flag when Face ID skipped
We need to store a flag for an edge case concerning a returning user.
E.g if user A has enabled biometrics but user B skips then previously
the code wouldn't respect the skip and try to prompt for biometrics. To
ensure that this flag can be removed, we then set the flag to false when
a user attempts to access Face ID settings.

A slight hack but the skip is on local auth onboarding is now respected.