[GOVUKAPP-1865] Feature/biometrics settings (#412)
* Remove setting passcode implicity

Previously, passcode was set when a user didn't have faceid configured
but had a passcode without notifiying the user, now we will force user
to manually login if faceid / touchid not configured.

* Removes local auth flag

We're going to be relying on device settings, vs maintaining an internal
state for local auth, at least for faceid which will simplify the
management of biometrics.

* Add skeleton settings functionality for faceid

touchid follows a different pattern and will be added.

* Ensure device can auth by biometrics for onboarding

* Adds logic to either return faceid or touchid

Also doens't display the setting if the device is capable of touchid /
faceid but has not set it up at the device level.

Settings will show if:
- faceid is enrolled (toggled on for app in device)
- faceid is disallowed (toggled off for app in device)

Biometrics enrolement will show if user has never allowed / disallowed
biometrics (skipped during onboarding).

* Update content and faceid view to match designs

* Add alert before opening settings for biometrics

* Add touchid settings

Sets a local flag to manage whether touchid is enabled. If it is skipped
then toggling the setting will attempt to enroll the user.

* Fixup tests

* Add tests for LocalAuthenticationSettingsViewModel

* Handle specific not available error

If user cancels enrollement then don't take to settings

* Update content to 30 days from 6 hours

* Add tracking

* Add more tests

* Update existing snapshot tests

* Throws error when full suite ran, not in isolation

XCTest/HarnessEventHandler.swift:182: Fatal error: Issue ‘Expectation
failed: (sut →
govuk_ios.LocalAuthenticationSettingsViewModel).showSettingsAlert →
false’ was recorded at
LocalAuthenticationSettingsViewModelTests.swift:118:13 without an
associated test. This may be due to calling #expect() or Issue.record()
from a detached Task or another unknown context.

* Add snapshot tests for biometrics views

* Fix test

* Fix sonar + pr comments

* Adds flag when Face ID skipped

We need to store a flag for an edge case concerning a returning user.
E.g if user A has enabled biometrics but user B skips then previously
the code wouldn't respect the skip and try to prompt for biometrics. To
ensure that this flag can be removed, we then set the flag to false when
a user attempts to access Face ID settings.

A slight hack but the skip is on local auth onboarding is now respected.

---------

Co-authored-by: Joshua Dubey <jodubey@deloitte.co.uk>