Prefix flag names with test
Uses test flag names which prevents the chat feature being accidentally
switched on for non produciton versions of chat (if app is deployed to
live before feature complete) when the real switches are enabled.

Having test flags is safe and allows us to write tests and potentially
test in real terms by adding them to the integration version of remote
config.